CRUDs Atividade, Demmanda e Grupo.

+++ apontamento/apontamento-service/src/main/java/br/com/porto/controlesinternos/apontamento/service/DemandaService.java
@@ -10,11 +10,11 @@ public interface DemandaService {
 
 	public boolean inserir(Demanda demanda);
 
-	public Demanda selecionarPorCodigo(long codigo);
+	public Demanda selecionarPorCodigo(int codigo);
 
 	public void alterar(Demanda demanda);
 
-	public void deletar(long codigo);
+	public void deletar(int codigo);
 
 	public List<Demanda> listar();
 }
